Refactor apps into top-level
