+++ astoria/astctl/usercode/kill.py
@@ -4,8 +4,8 @@ from typing import Optional
 
 import click
 
+from astoria.astctl.command import Command
 from astoria.common.ipc import ManagerRequest
-from astoria.consumers.astctl.command import Command
 
 loop = asyncio.get_event_loop()
