+++ astoria/astctl/usercode/show.py
@@ -4,8 +4,8 @@ from typing import Optional
 
 import click
 
+from astoria.astctl.command import SingleManagerMessageCommand
 from astoria.common.ipc import ProcessManagerMessage
-from astoria.consumers.astctl.command import SingleManagerMessageCommand
 
 loop = asyncio.get_event_loop()
