+++ astoria/astdiskd/static.py
