--- astoria/consumers/__init__.py
@@ -1,1 +1,0 @@
-"""State consumers."""
