--- astoria/managers/mixins/__init__.py
@@ -1,1 +1,0 @@
-"""Manager Mixins."""
