+++ tests/consumers/astwifid/test_wifi_hotspot_lifecycle.py
@@ -6,9 +6,9 @@ from typing import List, Tuple
 
 import pytest
 
+from astoria.astwifid import WiFiHotspotLifeCycle
 from astoria.common.config import AstoriaConfig
 from astoria.common.metadata import Metadata
-from astoria.consumers.astwifid import WiFiHotspotLifeCycle
 
 
 class FakeHostapdWiFiHotspotLifeCycle(WiFiHotspotLifeCycle):
