+++ tests/managers/mixins/test_disk_handler.py
@@ -8,7 +8,7 @@ import pytest
 
 from astoria.common.disks import DiskInfo, DiskType, DiskUUID
 from astoria.common.ipc import DiskManagerMessage
-from astoria.managers.mixins.disk_handler import DiskHandlerMixin
+from astoria.common.mixins import DiskHandlerMixin
 
 
 def get_match() -> Match[str]:
